Remove gyblets from FloatingPoint.swift
We only use .gyb features in two places in the FloatingPoint protocols, one of which we should be able to remove entirely: concrete inits from the builtin integer types, since those should be obsoleted by the BinaryInteger inits.

The other use is for random(in:), where it's only used to abstract over [Closed]Range. This mostly saves us some comment duplication, as the actually implementations are substantially different.

Also split BinaryFloatingPoint out into a new file, to make these files somewhat more manageable.
