+++ stdlib/public/core/CMakeLists.txt
@@ -60,7 +60,10 @@ set(SWIFTLIB_ESSENTIAL
   FixedArray.swift.gyb
   FlatMap.swift
   Flatten.swift
-  FloatingPoint.swift.gyb
+  FloatingPoint.swift
+  # ORDER DEPENDENCY: must follow CompilerProtocols.swift, may as well keep
+  # grouped together with the other FloatingPoint things.
+  BinaryFloatingPoint.swift
   FloatingPointParsing.swift.gyb
   FloatingPointTypes.swift.gyb
   Hashable.swift
