+++ stdlib/public/core/GroupInfo.json
@@ -139,6 +139,7 @@
       "Integers.swift",
       "IntegerParsing.swift"],
     "Floating": [
+      "BinaryFloatingPoint.swift",
       "FloatingPoint.swift",
       "FloatingPointParsing.swift",
       "FloatingPointTypes.swift"]}
